Fix Placeholder translateLabel
