+++ packages/forms/src/Components/Placeholder.php
@@ -2,11 +2,9 @@
 
 namespace Filament\Forms\Components;
 
-use Illuminate\Contracts\Support\Htmlable;
-use Illuminate\Support\Str;
-
 class Placeholder extends Component
 {
+    use Concerns\HasFallbackLabel;
     use Concerns\HasHelperText;
     use Concerns\HasHint;
     use Concerns\HasName;
@@ -51,14 +49,6 @@ class Placeholder extends Component
     public function getId(): string
     {
         return parent::getId() ?? $this->getStatePath();
-    }
-
-    public function getLabel(): string | Htmlable | null
-    {
-        return parent::getLabel() ?? (string) Str::of($this->getName())
-            ->kebab()
-            ->replace(['-', '_'], ' ')
-            ->ucfirst();
     }
 
     public function getContent()
